pyproject.toml
[project]
name = "3dprinter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.19.0",
    "openai-agents>=0.4.2",
]
